NDS Castlevania RetroArch 'cheat search' codes
Intended as a minimal alternative for the unwieldly enormous cheat files with hundreds of cheats that are difficult to scroll through.  These new (RetroArch) ones focus on the most pivotal things that a player might want to occasionally cheat for convenience, like health via potions or credits/continues in a game with limited continues.